Dry Run option added for tag removal tool

+++ bin/docker_utils.py
@@ -36,14 +36,15 @@ def get_dockerHubToken():
   response = http_request(url, data, None, method='POST')
   return response['token']
 
-def deleteTag(token, repo, tag):
-  print('Deleting tag: %s from %s repository....'% (tag, repo))
-  url = '%srepositories/%s/tags/%s/' % (DOCKER_HUB_API, repo, tag)
-  headers = {}
-  headers['Accept'] = 'application/json'
-  headers['Authorization'] = 'JWT %s' % token
-  response = http_request(url, None, headers, method = 'DELETE')
-  return response
+def deleteTag(token, repo, tag, dryRun):
+  print('** Deleting tag: %s from %s repository....'% (tag, repo))
+  if not dryRun:
+    url = '%srepositories/%s/tags/%s/' % (DOCKER_HUB_API, repo, tag)
+    headers = {}
+    headers['Accept'] = 'application/json'
+    headers['Authorization'] = 'JWT %s' % token
+    response = http_request(url, None, headers, method = 'DELETE')
+    return response
 
 def logout(token):
   url = '%slogout/' % DOCKER_HUB_API
@@ -63,7 +64,6 @@ def get_manifest(image):
   token = get_docker_token(repo)
   headers = {}
   headers['Authorization'] = 'Bearer %s' % token
-  print('Getting image_manifest for %s:%s' % (repo, tag))
   response = http_request(url, None, headers)
   return response
